moving create private channel intent to separate file
